# solutions/rsw-reference/RSW.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RSW Official",
    "version": "1.0.0",
    "contact": {}
  },
  "paths": {
...
    "/environmentTypes/{envTypeId}/configurations": {
      "get": {
        "tags": ["envTypeConfig"],
        "summary": "List envTypeConfigs",
        "description": "List envTypeConfigs",
        "operationId": "listEnvtypeconfigs",
        "parameters": [
          {
            "name": "paginationToken",
            "in": "query",
            "schema": {
              "type": "string",
              "example": ""
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "post": {
        "tags": ["envTypeConfig"],
        "summary": "Create envTypeConfig",
        "description": "Create envTypeConfig",
        "operationId": "createEnvtypeconfig",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "example": "Example config 1"
                  },
                  "estimatedCost": {
                    "type": "string",
                    "example": "estimated cost"
                  },
                  "name": {
                    "type": "string",
                    "example": "config 1"
                  },
                  "params": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "param1"
                    },
                    "example": []
                  },
                  "type": {
                    "type": "string",
                    "example": "sagemakerNotebook"
                  }
                }
              },
              "example": {
                "description": "Example config 1",
                "estimatedCost": "estimated cost",
                "name": "config 1",
                "params": [],
                "type": "sagemakerNotebook"
              }
            }
          }
        },
        "responses": {
          "201": {}
        }
      },
      "parameters": [
        {
          "name": "envTypeId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          },
          "description": "ID of environment type associated with this config"
        }
      ]
    },
    "/environmentTypes/{envTypeId}/configurations/{envTypeConfigId}": {
      "get": {
        "tags": ["envTypeConfig"],
        "summary": "Get envTypeConfig",
        "description": "Get envTypeConfig",
        "operationId": "getEnvtypeconfig",
        "parameters": [
          {
            "name": "paginationToken",
            "in": "query",
            "schema": {
              "type": "string",
              "example": ""
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "delete": {
        "tags": ["envTypeConfig"],
        "summary": "Soft Delete envTypeConfig",
        "description": "Soft Delete envTypeConfig",
        "operationId": "softDeleteEnvtypeconfig",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "204": {}
        }
      },
      "patch": {
        "tags": ["envTypeConfig"],
        "summary": "Update envTypeConfig",
        "description": "Update envTypeConfig",
        "operationId": "updateEnvtypeconfig",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "example": "Example config 1"
                  },
                  "estimatedCost": {
                    "type": "string",
                    "example": "estimatedCost"
                  }
                }
              },
              "example": {
                "description": "Example config 1",
                "estimatedCost": "estimatedCost"
              }
            }
          }
        },
        "responses": {
          "200": {}
        }
      },
      "parameters": [
        {
          "name": "envTypeId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          },
          "description": "ID of environment type associated with this config"
        },
        {
          "name": "envTypeConfigId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          },
          "description": "ID of environment type config to associate"
        }
      ]
    },
...
  },
...
}